Change local Storage for Leaderboard API storage
