--- src/modules/new-score.js
@@ -1,12 +1,0 @@
-// Functions to create a new score object
-
-class NewScore {
-
-    constructor(name, score) {
-        this.name = name;
-        this.score = score;
-    }
-
-}
-
-export default NewScore
