--- src/modules/refresh.js
@@ -1,11 +1,0 @@
-import displayScores from "./display-scores"
-import LocalStorage from "./Local-storage"
-
-const refresh = () => {
-  let recentScores = LocalStorage.getScores()
-  recentScores = []
-  displayScores(recentScores);
-  localStorage.setItem('recentScores', JSON.stringify(recentScores));
-}
-
-export default refresh
